requirements.txt
numpy
mcp
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pytest
google-cloud-secret-manager
google-genai
python-dotenv
arxiv
requests
PyYAML>=6.0.1
pytest-asyncio>=0.21.0
